+++ NEWS.md
@@ -1,3 +1,10 @@
 # wgsLR 0.0.1.9001
 
 * Added a `NEWS.md` file to track changes to the package.
+* Renamed objects:
+    - `d_formulas_Hp` renamed to `d_formulas_Hp_w`
+    - `d_prob_Hp` renamed to `d_prob_Hp_w`
+    - `d_formulas_Hd` renamed to `d_formulas_Hd_w`
+    - `d_formulas_Hp` renamed to `d_formulas_Hp_w`
+* Added objects related to sample specific error probabilities: 
+    - `d_formulas_Hp_wDwS`